Provide counters for each state
